+++ apps/heureka/src/components/Services/ServicesList/index.test.tsx
@@ -16,10 +16,31 @@ const renderServicesList = () => ({
     <TestProvider>
       <MessagesProvider>
         <ServicesList
-          filterSettings={{
-            searchTerm: "",
-            selectedFilters: [],
-          }}
+          services={[
+            {
+              id: "1",
+              name: "alpha",
+              issuesCount: {
+                critical: 0,
+                high: 0,
+                medium: 0,
+                low: 0,
+                none: 0,
+                total: 0,
+              },
+              serviceDetails: {
+                supportGroups: [],
+              },
+              components: 0,
+              remediationDate: "",
+              serviceOwners: [],
+            },
+          ]}
+          loading={false}
+          error={""}
+          currentPage={1}
+          totalNumberOfPages={1}
+          goToPage={() => {}}
         />
       </MessagesProvider>
     </TestProvider>
@@ -32,7 +53,7 @@ describe("ServicesList", () => {
 
     // Check for the presence of the service list headers
     expect(await screen.findByText("Service")).toBeInTheDocument()
-    expect(await screen.findByText("Service details")).toBeInTheDocument()
+    expect(await screen.findByText("Details")).toBeInTheDocument()
   })
 
   it("should render service panel when clicking on a service", async () => {