+++ packages/k8s-client/vite.config.ts
@@ -3,16 +3,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { defineConfig } from "vite"
 import dts from "vite-plugin-dts"
 
-export default defineConfig({
+export default {
   build: {
     lib: {
       entry: "src/index.ts", // or 'src/main.ts' if TypeScript
       name: "k8s-client", // Replace with your library's global name
-      formats: ["es"], // Output formats: ESM and CommonJS
-      fileName: () => `index.js`, // Output file names
+      formats: ["es", "cjs"], // Output formats: ESM and CommonJS
+      fileName: (format) => `index.${format}.js`, // Output file names
     },
     outDir: "build",
   },
@@ -23,4 +22,4 @@ export default defineConfig({
       outDir: "build", // Specify where to output the types
     }),
   ],
-})
+}