+++ packages/ui-components/src/components/TooltipTrigger/TooltipTrigger.component.tsx
@@ -31,12 +31,13 @@ export const TooltipTrigger = React.forwardRef<HTMLElement, TooltipTriggerProps>
   const state = useTooltipState()
 
   // merge all the refs
-  const ref = useMergeRefs([state.refs.setReference, propRef])
+  const ref = useMergeRefs([state.refs?.setReference, propRef])
   // `asChild` allows the user to pass any element as the anchor
   if (asChild && React.isValidElement(children)) {
     const childrenProps = children.props as CustomProps
     const childrenRef = childrenProps.ref as React.RefObject<HTMLElement>
-    const ref = useMergeRefs([state.refs.setReference, propRef, childrenRef])
+    const setReference = state.refs?.setReference
+    const ref = useMergeRefs([setReference, propRef, childrenRef])
 
     const referencedProps = {
       ref,