+++ apps/heureka/CHANGELOG.md
@@ -1,5 +1,47 @@
 # heureka
 
+## 0.5.4
+
+### Patch Changes
+
+- 43d6aa0: Update vite to the latest version (v 6.3.5) to resolve dependency vulnarebilities
+- Updated dependencies [cdae407]
+- Updated dependencies [7230774]
+- Updated dependencies [a9e8592]
+- Updated dependencies [43d6aa0]
+  - @cloudoperators/juno-ui-components@3.1.1
+  - @cloudoperators/juno-messages-provider@0.2.5
+
+## 0.5.3
+
+### Patch Changes
+
+- 3140768: Refactor Issue Count Handling and Add Total Counts per Service
+
+## 0.5.2
+
+### Patch Changes
+
+- 37e1bf4: Added a search bar to the issues list that allows users to filter issues by CVE name (primaryName).
+- 403b5cd: - Remove headlessui float and update headlessui to 2.1.10
+  - Fix unit tests
+- Updated dependencies [403b5cd]
+- Updated dependencies [67f53cb]
+  - @cloudoperators/juno-ui-components@3.1.0
+
+## 0.5.1
+
+### Patch Changes
+
+- 52e148c: Refactors severity count related components
+- ee1b3ab: Reorganize components files based on how they are imported.
+- d50e279: Improve issue description truncation in ImageVersionIssueListItem
+- 0d6c4e8: Closes image version details panel on reselect in Service Details page
+- Updated dependencies [9ff1784]
+- Updated dependencies [4e8a5ef]
+- Updated dependencies [f48439c]
+  - @cloudoperators/juno-ui-components@3.0.1
+
 ## 0.5.0
 
 ### Minor Changes