+++ packages/k8s-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cloudoperators/juno-k8s-client
 
+## 1.0.6
+
+### Patch Changes
+
+- 43d6aa0: Update vite to the latest version (v 6.3.5) to resolve dependency vulnarebilities
+
+## 1.0.5
+
+### Patch Changes
+
+- f3be4a6: add support for common js
+
 ## 1.0.4
 
 ### Patch Changes